Add full typescript support + eslint
