+++ .eslintrc.json
@@ -1,0 +1,21 @@
+{
+  "extends": [
+    "eslint:recommended",
+    "prettier",
+    "prettier/standard"
+  ],
+  "parser": "typescript-eslint-parser",
+  "env": {
+    "node": true,
+    "es6": true
+  },
+  "parserOptions": {
+    "sourceType": "module",
+    "ecmaFeatures": {
+      "modules": true
+    }
+  },
+  "rules": {
+    "no-console": 0
+  }
+}
