+++ .gitignore
@@ -6,5 +6,6 @@
 
 # misc
 .DS_Store
+.vscode
 
 npm-debug.log*
