+++ .prettierrc
@@ -1,4 +1,5 @@
 {
     "singleQuote": true,
+    "semi": true,
     "jsxSingleQuote": true
 }
