+++ package.json
@@ -4,6 +4,8 @@
   "description": "Find out the cost of webpack bundle over time",
   "main": "index.js",
   "scripts": {
+    "build": "tsc",
+    "lint:plugin": "eslint 'plugin/**/*.ts'",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -19,6 +21,15 @@
   "dependencies": {
     "chalk": "^2.4.1",
     "firebase": "^5.5.8",
-    "gzip-size": "^5.0.0"
+    "gzip-size": "^5.0.0",
+    "lodash": "^4.17.11",
+    "typescript": "^3.1.6"
+  },
+  "devDependencies": {
+    "@types/gzip-size": "^4.1.0",
+    "@types/lodash": "^4.14.118",
+    "eslint": "^5.9.0",
+    "eslint-config-prettier": "^3.3.0",
+    "typescript-eslint-parser": "^21.0.1"
   }
 }
